refactor(frontend): Add lazy-loaded solution module
