+++ frontend/src/app/assignment/assignment-routing.module.ts
@@ -1,6 +1,5 @@
 import {NgModule} from '@angular/core';
-import {Route, RouterModule, Routes} from '@angular/router';
-import {CommentListComponent} from './components/comment-list/comment-list.component';
+import {RouterModule, Routes} from '@angular/router';
 
 import {AssignmentComponent} from './pages/assignment/assignment.component';
 import {ShareComponent} from './pages/assignment/share/share.component';
@@ -18,14 +17,9 @@ import {SampleComponent} from './pages/edit-assignment/sample/sample.component';
 import {TasksComponent} from './pages/edit-assignment/tasks/tasks.component';
 import {TemplateComponent} from './pages/edit-assignment/template/template.component';
 import {EditTaskModalComponent} from './pages/edit-task-modal/edit-task-modal.component';
-import {EvaluationModalComponent} from './pages/evaluation-modal/evaluation-modal.component';
 import {MyAssignmentsComponent} from './pages/my-assignments/my-assignments.component';
 import {MyCoursesComponent} from './pages/my-courses/my-courses.component';
 import {MySolutionsComponent} from './pages/my-solutions/my-solutions.component';
-import {SolutionDetailsComponent} from './pages/solution/info/details.component';
-import {SolutionShareComponent} from './pages/solution/share/share.component';
-import {SolutionComponent} from './pages/solution/solution.component';
-import {SolutionTasksComponent} from './pages/solution/tasks/tasks.component';
 import {TokenModalComponent} from './pages/token-modal/token-modal.component';
 
 export const editAssignmentChildRoutes: Routes = [
@@ -60,16 +54,6 @@ export const assignmentChildRoutes = [
   {path: 'token', component: TokenModalComponent},
 ];
 
-export const solutionChildRoutes: Routes = [
-  {
-    path: 'tasks', component: SolutionTasksComponent, data: {title: 'Solution & Tasks'}, children: [
-      {path: ':task', component: EvaluationModalComponent},
-    ],
-  },
-  {path: 'details', component: SolutionDetailsComponent, data: {title: 'Student Info'}},
-  {path: 'share', component: SolutionShareComponent, data: {title: 'Sharing'}},
-  {path: 'comments', component: CommentListComponent, data: {title: 'Comments'}},
-];
 
 const routes: Routes = [
   {path: '', component: MyAssignmentsComponent},
@@ -86,12 +70,7 @@ const routes: Routes = [
   {path: ':aid/solutions/new', component: CreateSolutionComponent},
   {
     path: ':aid/solutions/:sid',
-    component: SolutionComponent,
-    children: [
-      ...solutionChildRoutes,
-      {path: 'token', component: TokenModalComponent},
-      {path: '', redirectTo: 'tasks'},
-    ],
+    loadChildren: () => import('./modules/solution/solution.module').then(s => s.SolutionModule),
   },
   {
     path: ':aid/edit',
