+++ frontend/src/app/assignment/modules/solution/details/details.component.spec.ts
