+++ frontend/src/app/assignment/modules/solution/evaluation-modal/evaluation-modal.component.html
