+++ frontend/src/app/assignment/modules/solution/share/share.component.html
