+++ frontend/src/app/assignment/modules/solution/solution/solution.component.scss
