+++ frontend/src/app/assignment/modules/solution/solution/solution.component.ts
@@ -2,11 +2,11 @@ import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
 import {combineLatest, forkJoin} from 'rxjs';
 import {switchMap} from 'rxjs/operators';
-import {solutionChildRoutes} from '../../assignment-routing.module';
-import Assignment from '../../model/assignment';
-import Solution from '../../model/solution';
-import {AssignmentService} from '../../services/assignment.service';
-import {SolutionService} from '../../services/solution.service';
+import Assignment from '../../../model/assignment';
+import Solution from '../../../model/solution';
+import {AssignmentService} from '../../../services/assignment.service';
+import {SolutionService} from '../../../services/solution.service';
+import {solutionChildRoutes} from '../solution-routing.module';
 
 @Component({
   selector: 'app-solution',
